+++ tests/test_io.py
@@ -8,7 +8,7 @@ import numpy as np
 
 import aspecd.exceptions
 import aspecd.processing
-from aspecd import io, dataset, tasks, utils
+from aspecd import io, dataset, tasks, utils, history
 
 
 class TestDatasetImporter(unittest.TestCase):
@@ -91,6 +91,16 @@ class TestDatasetExporter(unittest.TestCase):
         test_dataset = dataset.Dataset()
         self.exporter.export_from(test_dataset)
         self.assertIs(self.exporter.dataset, test_dataset)
+
+    def test_has_create_history_record_method(self):
+        self.assertTrue(hasattr(self.exporter, 'create_history_record'))
+        self.assertTrue(callable(self.exporter.create_history_record))
+
+    def test_create_history_record_returns_history_record(self):
+        self.exporter.dataset = aspecd.dataset.Dataset()
+        history_record = self.exporter.create_history_record()
+        self.assertTrue(isinstance(history_record,
+                                   aspecd.history.DatasetExporterHistoryRecord))
 
 
 class TestDatasetImporterFactory(unittest.TestCase):